Fixed bug for Smart easy fills of squares
The Smart way of solving a Sodoku puzzle relies on context from the puzzle.  One of the criteria is an "easy fill" which is a row, column, or square which only has one number missing, and so the number can be determined.  The bug had taken the coordinates of the filled in tile which was calling the easyFills method to check the relevant row, column, and square and instead of recursively calling easyFills with the newest coordinate of a tile that had been easily filled, it was calling it with the original set of coordinates (ie nothing new was happening, it was just recursively calling itself without a point).